Clean up unused classes and tests
