--- src/Manager/Factory.php
@@ -1,113 +1,0 @@
-<?php
-
-namespace Dkan\Datastore\Manager;
-
-use Dkan\Datastore\LockableBinStorage;
-use Dkan\Datastore\Resource;
-use Dkan\Datastore\Storage\IDatabase;
-
-/**
- * Class Factory.
- *
- * Builds a datastore manager for a resource.
- */
-class Factory {
-  private $binStorage;
-  private $database;
-
-  private $resource;
-  private $provider;
-
-  private $classes = [];
-
-  /**
-   * Constructor.
-   */
-  public function __construct(Resource $resource, InfoProvider $provider, LockableBinStorage $bin_storage, IDatabase $database) {
-    $this->binStorage = $bin_storage;
-    $this->resource = $resource;
-    $this->provider = $provider;
-    $this->database = $database;
-    $this->setupClassHierarchy();
-  }
-
-  /**
-   * Set a class.
-   *
-   * This class will be given priority when trying to
-   * create a datastore manager for the given resource.
-   */
-  public function setClass($class) {
-    array_unshift($this->classes, $class);
-  }
-
-  /**
-   * Get the datastore manager.
-   */
-  public function get() {
-    foreach ($this->classes as $class) {
-      try {
-        return $this->getManager($class);
-      }
-      catch (\Exception $e) {
-      }
-    }
-    throw new \Exception("Datastore could not be loaded");
-
-  }
-
-  /**
-   * Get manager.
-   */
-  private function getManager($class) {
-    $exists = class_exists($class);
-    if (!$exists) {
-      throw new \Exception("The class {$class} does not exist.");
-    }
-
-    $interfaces = class_implements($class);
-    $interface = "Dkan\Datastore\Manager\IManager";
-    if (!in_array($interface, $interfaces)) {
-      throw new \Exception("The class {$class} does not implement the interface {$interface}.");
-    }
-
-    return new $class($this->resource, $this->binStorage, $this->database);
-  }
-
-  /**
-   * Create an array of datastore manager classes.
-   *
-   * It picks a "random" manager class, and, if available,
-   * the class from the datastore state.
-   *
-   * This is useful as a hierarcy to eliminate failures
-   * if our datastore state becomes corrupt, and the official
-   * manager is no longer valid or available.
-   */
-  private function setupClassHierarchy() {
-    array_unshift($this->classes, $this->getClass());
-
-    $state = $this->binStorage->get($this->resource->getId());
-
-    if ($state) {
-      $class = $state['class'];
-      array_unshift($this->classes, $class);
-    }
-  }
-
-  /**
-   * Choose the first class from the manager's info array.
-   */
-  private function getClass() {
-    $info = $this->provider->getInfo();
-
-    if (empty($info)) {
-      throw new \Exception("No Datastore Managers are active");
-    }
-
-    /* @var $i \Dkan\Datastore\Manager\Info */
-    $i = array_shift($info);
-    return $i->getClass();
-  }
-
-}
