+++ test/Manager/SimpleImport/SimpleImportTest.php
@@ -1,10 +1,10 @@
 <?php
 
-use Dkan\Datastore\Manager\SimpleImport\SimpleImport;
+use Contracts\Mock\Storage\Memory;
 use Dkan\Datastore\Resource;
 use Locker\Locker;
 
-class SimpleImportTest extends \PHPUnit\Framework\TestCase {
+class SimpleImport2Test extends \PHPUnit\Framework\TestCase {
 
   private $database;
 
@@ -13,49 +13,35 @@ class SimpleImportTest extends \PHPUnit\Framework\TestCase {
    */
   protected function setUp()
   {
-    $this->database = new \Dkan\Datastore\Storage\Database\Memory();
+    $this->database = new TestMemStorage2();
   }
 
   private function getDatastore(Resource $resource) {
-
-    $provider = new \Dkan\Datastore\Manager\InfoProvider();
-    $provider->addInfo(new \Dkan\Datastore\Manager\Info(SimpleImport::class, "simple_import", "SimpleImport"));
-
-    $bin_storage = new \Dkan\Datastore\LockableBinStorage("dkan_datastore", new Locker("dkan_datastore"), new \Dkan\Datastore\Storage\KeyValue\Memory());
-
-    $factory = new \Dkan\Datastore\Manager\Factory($resource, $provider, $bin_storage, $this->database);
-
-    return $factory->get();
+    return new \Dkan\Datastore\Manager($resource, $this->database, \CsvParser\Parser\Csv::getParser());
   }
 
   public function testBasics() {
     $resource = new Resource(1, __DIR__ . "/../../data/countries.csv");
 
-    /* @var $datastore \Dkan\Datastore\Manager\SimpleImport\SimpleImport */
     $datastore = $this->getDatastore($resource);
 
     $status = $datastore->getStatus();
-    $this->assertEquals(SimpleImport::STORAGE_UNINITIALIZED, $status['storage']);
-    $this->assertEquals(SimpleImport::DATA_IMPORT_UNINITIALIZED, $status['data_import']);
+    $this->assertEquals($datastore::DATA_IMPORT_UNINITIALIZED, $status['data_import']);
 
     $datastore->import();
 
     $status = $datastore->getStatus();
-    $this->assertEquals(SimpleImport::STORAGE_INITIALIZED, $status['storage']);
-    $this->assertEquals(SimpleImport::DATA_IMPORT_DONE, $status['data_import']);
+    $this->assertEquals($datastore::DATA_IMPORT_DONE, $status['data_import']);
 
-    $this->assertEquals(4, $datastore->numberOfRecordsImported());
+    $this->assertEquals(4, $datastore->getStorage()->count());
 
     $status = $datastore->getStatus();
-    $this->assertEquals(SimpleImport::STORAGE_INITIALIZED, $status['storage']);
-    $this->assertEquals(SimpleImport::DATA_IMPORT_DONE, $status['data_import']);
+    $this->assertEquals($datastore::DATA_IMPORT_DONE, $status['data_import']);
 
     $datastore->drop();
-    $this->assertFalse($this->database->tableExist($datastore->getTableName()));
 
     $status = $datastore->getStatus();
-    $this->assertEquals(SimpleImport::STORAGE_UNINITIALIZED, $status['storage']);
-    $this->assertEquals(SimpleImport::DATA_IMPORT_UNINITIALIZED, $status['data_import']);
+    $this->assertEquals($datastore::DATA_IMPORT_UNINITIALIZED, $status['data_import']);
   }
 
   public function testOver1000() {
@@ -64,13 +50,22 @@ class SimpleImportTest extends \PHPUnit\Framework\TestCase {
     $datastore = $this->getDatastore($resource);
     $datastore->import();
 
-    $this->assertEquals(2969, $datastore->numberOfRecordsImported());
+    $this->assertEquals(2969, $datastore->getStorage()->count());
+
+    $results = $datastore->getStorage()->retrieveAll();
+    $values = array_values($results);
 
     $expected = '["2049","75000403","R","1","DESIGNATED","0.076","0.364","463.2487"]';
-    $this->assertEquals($expected, $this->database->tables['dkan_datastore_1'][0]);
+    $this->assertEquals($expected, $values[0]);
 
     $expected = '["2048","75000402","R","1","DESIGNATED","0.769","1.713","1528.0913"]';
-    $this->assertEquals($expected, $this->database->tables['dkan_datastore_1'][2968]);
+    $this->assertEquals($expected, $values[2968]);
   }
+}
 
+class TestMemStorage2 extends Contracts\Mock\Storage\Memory implements \Dkan\Datastore\Storage\Storage {
+  public function count(): int
+  {
+    return count($this->storage);
+  }
 }
